agent/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
psutil==5.9.8
mac-vendor-lookup==0.1.12
mss==9.0.1
Pillow>=10.0.0
